flash_card/requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
requests==2.31.0
python-dotenv==1.0.0
torch
diffusers
transformers
Pillow==10.2.0 